Refactor plugin names for consistency and accuracy
Use `splitgraph-` prefix where applicable, and be more accurate
about export destination, e.g. `SplitgraphExportQueryToFilePlugin`
instead of `ExportQueryPlugin`
