+++ packages/react/hooks.test.tsx
@@ -227,8 +227,8 @@ describe("makeDefaultAnonymousContext", () => {
                 },
                 "transformRequestHeaders": [Function],
               },
-              ExportQueryPlugin {
-                "__name": "exportQuery",
+              SplitgraphExportQueryToFilePlugin {
+                "__name": "export-query-to-file",
                 "graphqlClient": SplitgraphGraphQLClient {
                   "graphqlClient": GraphQLClient {
                     "options": {
@@ -270,8 +270,8 @@ describe("makeDefaultAnonymousContext", () => {
                 },
                 "transformRequestHeaders": [Function],
               },
-              ExportQueryPlugin {
-                "__name": "exportQuery",
+              SplitgraphExportQueryToFilePlugin {
+                "__name": "export-query-to-file",
                 "graphqlClient": SplitgraphGraphQLClient {
                   "graphqlClient": GraphQLClient {
                     "options": {
